requirements.txt
openai
streamlit-nightly>=1.27.3.dev20231006